+++ utils/vpc/projectgenerator_xbox360_2010.cpp
@@ -105,7 +105,7 @@ bool CProjectGenerator_Xbox360_2010::WriteFile(CProjectFile *pFile,
     m_XMLWriter.PushNode(pKeyName,
                          CFmtStr("Include=\"%s\"", pFile->m_Name.Get()));
 
-    for (int i = 0; i < pFile->m_Configs.Count(); i++) {
+    for (intp i = 0; i < pFile->m_Configs.Count(); i++) {
       if (!WriteConfiguration(pFile->m_Configs[i])) return false;
     }
 
@@ -235,7 +235,7 @@ bool CProjectGenerator_Xbox360_2010::WritePrimaryXML(
   m_XMLWriter.PushNode("ItemGroup", "Label=\"ProjectConfigurations\"");
   CUtlVector<CUtlString> configurationNames;
   m_pVCProjGenerator->GetAllConfigurationNames(configurationNames);
-  for (int i = 0; i < configurationNames.Count(); i++) {
+  for (intp i = 0; i < configurationNames.Count(); i++) {
     m_XMLWriter.PushNode(
         "ProjectConfiguration",
         CFmtStr("Include=\"%s|Xbox 360\"", configurationNames[i].Get()));
@@ -256,7 +256,7 @@ bool CProjectGenerator_Xbox360_2010::WritePrimaryXML(
       "<Import Project=\"$(VCTargetsPath)\\Microsoft.Cpp.Default.props\" />");
 
   // write the root configurations
-  for (int i = 0; i < configurationNames.Count(); i++) {
+  for (intp i = 0; i < configurationNames.Count(); i++) {
     CProjectConfiguration *pConfiguration = NULL;
     if (m_pVCProjGenerator->GetRootConfiguration(configurationNames[i].Get(),
                                                  &pConfiguration)) {
@@ -269,7 +269,7 @@ bool CProjectGenerator_Xbox360_2010::WritePrimaryXML(
   m_XMLWriter.PushNode("ImportGroup", "Label=\"ExtensionSettings\"");
   m_XMLWriter.PopNode(true);
 
-  for (int i = 0; i < configurationNames.Count(); i++) {
+  for (intp i = 0; i < configurationNames.Count(); i++) {
     m_XMLWriter.PushNode("ImportGroup",
                          CFmtStr("Condition=\"'$(Configuration)|$(Platform)'=='"
                                  "%s|Xbox 360'\" Label=\"PropertySheets\"",
@@ -286,7 +286,7 @@ bool CProjectGenerator_Xbox360_2010::WritePrimaryXML(
 
   m_XMLWriter.PushNode("PropertyGroup");
   m_XMLWriter.WriteLineNode("_ProjectFileVersion", "", "10.0.30319.1");
-  for (int i = 0; i < configurationNames.Count(); i++) {
+  for (intp i = 0; i < configurationNames.Count(); i++) {
     CProjectConfiguration *pConfiguration = NULL;
     if (m_pVCProjGenerator->GetRootConfiguration(configurationNames[i].Get(),
                                                  &pConfiguration)) {
@@ -338,7 +338,7 @@ bool CProjectGenerator_Xbox360_2010::WritePrimaryXML(
   m_XMLWriter.PopNode(true);
 
   // write the tool configurations
-  for (int i = 0; i < configurationNames.Count(); i++) {
+  for (intp i = 0; i < configurationNames.Count(); i++) {
     CProjectConfiguration *pConfiguration = NULL;
     if (m_pVCProjGenerator->GetRootConfiguration(configurationNames[i].Get(),
                                                  &pConfiguration)) {
